Refactor LearnMore
